qa-script-wpvs-free-master/requirements.txt
attrs
build
certifi
cffi
charset-normalizer
click
colorama
h11
idna
iniconfig
Jinja2
MarkupSafe
outcome
packaging
pip-review
pip-tools
pluggy
pycparser
pyproject_hooks
PySocks
pytest
pytest-html
pytest-metadata
pytest-order
python-dotenv
requests
selenium
setuptools
sniffio
sortedcontainers
trio
trio-websocket
typing_extensions
urllib3
webdriver-manager
websocket-client
wheel
wsproto
